finished community engagement slide + separated slide data
