+++ src/components/MeetTheTeam/slideData.js
@@ -1,0 +1,509 @@
+const slideData = {
+  coChairs: {
+    title: "Co-Chairs",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+      {
+        type: "image",
+        src: "src/assets/meettheteam/pink-flower.png",
+        position: "absolute top-[50vh] left-[2vw] w-[4.2vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/cake-square.png",
+        position: "absolute top-[35vh] right-[20vw] w-[15vw]"
+      },
+      {
+        type: "label",
+        text: "Co-Chairs",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      }
+    ],
+    members: [
+      {
+        name: "Julia Da Silva",
+        image: "src/assets/meettheteam/julia-nametag.png",
+        position: "absolute top-[5vh] left-[15vw] w-[16vw]"
+      },
+      {
+        name: "Katsa Bejemil", 
+        image: "src/assets/meettheteam/katsa-nametag.png",
+        position: "absolute top-[5vh] left-[35vw] w-[16vw]"
+      },
+      {
+        name: "Rumaisa Nawar",
+        image: "src/assets/meettheteam/rumaisa-nametag.png", 
+        position: "absolute top-[5vh] left-[55vw] w-[16vw]"
+      },
+      {
+        name: "Sabriha Amin",
+        image: "src/assets/meettheteam/sabriha-nametag.png",
+        position: "absolute top-[35vh] left-[25vw] w-[16vw]"
+      },
+      {
+        name: "Aditi Patel",
+        image: "src/assets/meettheteam/aditi-nametag.png",
+        position: "absolute top-[35vh] left-[45vw] w-[16vw]"
+      }
+    ]
+  },
+  Logistics1: {
+    title: "Logistics",
+    slideConfig: {
+      width: "100vw",
+      height: "68vh", 
+      bgColor: "#A7748E"
+    },
+    decorations: [
+      {
+        type: "label",
+        text: "Logistics",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/cake-rectangle.png",
+        position: "absolute top-[5vh] right-[10vw] w-[15vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[40vh] left-[2vw] w-[10vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/director-tag.png",
+        position: "absolute top-[-2vh] left-[29vw] w-[10vw]"
+      }
+    ],
+    members: [
+      {
+        name: "Triya Augustine",
+        image: "src/assets/meettheteam/triya-nametag.png",
+        position: "absolute top-[5vh] left-[20vw] w-[16vw]"
+      },
+      {
+        name: "Luiza Teles",
+        image: "src/assets/meettheteam/luiza-nametag.png", 
+        position: "absolute top-[5vh] left-[50vw] w-[6vw]"
+      },
+      {
+        name: "Vansh Madan",
+        image: "src/assets/meettheteam/vansh-nametag.png",
+        position: "absolute top-[34vh] left-[35vw] w-[15vw]"
+      },
+      {
+        name: "Anagha Koroth", 
+        image: "src/assets/meettheteam/anagha-nametag.png",
+        position: "absolute top-[34vh] left-[55vw] w-[16vw]"
+      }
+    ]
+  },
+  Logistics2: {
+    title: "Logistics",
+    slideConfig: {
+      width: "100vw",
+      height: "68vh", 
+      bgColor: "#A7748E"
+    },
+    decorations: [
+      {
+        type: "label",
+        text: "Logistics",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/green-flower.png", // brown flower
+        position: "absolute bottom-[2vh] left-[19vw] w-[5vw]"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/spoon.png", // spoon
+        position: "absolute top-[5vh] right-[30vw] w-[5vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/pie.png", // pink pie
+        position: "absolute top-[5vh] right-[21vw] w-[10vw]"
+      },
+    ],
+    members: [
+      {
+        name: "Seyedeh Sara Alav",
+        image: "src/assets/meettheteam/seyedeh-nametag.png",
+        position: "absolute top-[5vh] left-[20vw] w-[16vw]"
+      },
+      {
+        name: "Bohina Manochehrzadeh",
+        image: "src/assets/meettheteam/bohina-nametag.png", 
+        position: "absolute top-[5vh] left-[50vw] w-[6vw]"
+      },
+      {
+        name: "Irqa Inam",
+        image: "src/assets/meettheteam/iqra-nametag.png",
+        position: "absolute top-[34vh] left-[35vw] w-[15vw]"
+      },
+      {
+        name: "Maryam Malik",
+        image: "src/assets/meettheteam/maryam-nametag.png",
+        position: "absolute top-[34vh] left-[55vw] w-[16vw]"
+      }
+    ]
+  },
+  IT1: {
+    title: "IT",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+       {
+        type: "image", 
+        src: "src/assets/meettheteam/director-tag.png",
+        position: "absolute top-[-2vh] left-[27vw] w-[10vw]"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/pink-flower.png",
+        position: "absolute top-[50vh] left-[2vw] w-[4.2vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/cake-square.png",
+        position: "absolute top-[35vh] right-[20vw] w-[15vw]"
+      },
+      {
+        type: "label",
+        text: "IT",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      }
+    ],
+    members: [
+      {
+        name: "Joanne Wijetunga",
+        image: "src/assets/meettheteam/joanne-nametag.png",
+        position: "absolute top-[5vh] left-[18vw] w-[16vw]"
+      },
+      {
+        name: "Fares Islam", 
+        image: "src/assets/meettheteam/fares-nametag.png",
+        position: "absolute top-[5vh] left-[38vw] w-[16vw]"
+      },
+      {
+        name: "Yazan Haddad",
+        image: "src/assets/meettheteam/yazan-nametag.png",
+        position: "absolute top-[5vh] left-[58vw] w-[16vw]"
+      },
+      {
+        name: "Ifra Inam",
+        image: "src/assets/meettheteam/ifra-nametag.png",
+        position: "absolute top-[35vh] left-[28vw] w-[16vw]"
+      },
+      {
+        name: "Madeleine Phan",
+        image: "src/assets/meettheteam/madeleine-nametag.png",
+        position: "absolute top-[35vh] left-[48vw] w-[16vw]"
+      }
+    ]
+  },
+  IT2: {
+    title: "IT",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+       {
+        type: "image",
+        src: "src/assets/meettheteam/cake-rectangle.png",
+        position: "absolute top-[3vh] right-[41vw] w-[17vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[35vh] right-[27vw] w-[7vw] rotate-270"
+      },
+      {
+        type: "label",
+        text: "IT",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      }
+    ],
+    members: [
+      {
+        name: "Natalie Lewis",
+        image: "src/assets/meettheteam/natalie-nametag.png",
+        position: "absolute top-[5vh] left-[22vw] w-[16vw]"
+      },
+      {
+        name: "Nina Dang",
+        image: "src/assets/meettheteam/nina-nametag.png",
+        position: "absolute top-[5vh] left-[62vw] w-[16vw]"
+      },
+      {
+        name: "Angad Ahluwalia",
+        image: "src/assets/meettheteam/angad-nametag.png",
+        position: "absolute top-[35vh] left-[32vw] w-[16vw]"
+      },
+      {
+        name: "Sidney Liu",
+        image: "src/assets/meettheteam/sidney-nametag.png",
+        position: "absolute top-[35vh] left-[52vw] w-[16vw]"
+      }
+    ]
+  },
+  Marketing: {
+    title: "Marketing",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+       {
+        type: "image", 
+        src: "src/assets/meettheteam/pie.png", // pink pie
+        position: "absolute top-[45vh] right-[21vw] w-[10vw]"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/spoon.png", // spoon
+        position: "absolute top-[48vh] right-[18vw] w-[5vw] rotate-180"
+      },
+      {
+        type: "image",
+        src: "src/assets/meettheteam/pink-flower.png",
+        position: "absolute top-[50vh] left-[8vw] w-[4.2vw]"
+      },
+      {
+        type: "label",
+        text: "Marketing",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      }
+    ],
+    members: [
+      {
+        name: "Abeeha Zubair",
+        image: "src/assets/meettheteam/abeeha-nametag.png",
+        position: "absolute top-[5vh] left-[22vw] w-[16vw]"
+      },
+      {
+        name: "Julie Bautista",
+        image: "src/assets/meettheteam/julie-nametag.png",
+        position: "absolute top-[5vh] left-[42vw] w-[16vw]"
+      },
+      {
+        name: "Kimili Nguyen",
+        image: "src/assets/meettheteam/kimili-nametag.png",
+        position: "absolute top-[5vh] left-[62vw] w-[16vw]"
+      },
+      {
+        name: "Larissa Singh",
+        image: "src/assets/meettheteam/larissa-nametag.png",
+        position: "absolute top-[35vh] left-[32vw] w-[16vw]"
+      },
+      {
+        name: "Nabeela Ansari",
+        image: "src/assets/meettheteam/nabeela-nametag.png",
+        position: "absolute top-[35vh] left-[52vw] w-[16vw]"
+      }
+    ]
+  },
+  Design: {
+    title: "Design",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+       {
+        type: "image",
+        src: "src/assets/meettheteam/cake-rectangle.png",
+        position: "absolute top-[45vh] right-[42vw] w-[15vw] rotate-10"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[35vh] right-[27vw] w-[7vw] rotate-270"
+      },
+      {
+        type: "label",
+        text: "Design",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[5vh] left-[35vw] w-[7vw] rotate-270"
+      },
+    ],
+    members: [
+      {
+        name: "Samreen Ahmed",
+        image: "src/assets/meettheteam/samreen-nametag.png",
+        position: "absolute top-[35vh] left-[22vw] w-[16vw]"
+      },
+      {
+        name: "Subin Hwang",
+        image: "src/assets/meettheteam/subin-nametag.png",
+        position: "absolute top-[5vh] left-[41vw] w-[16vw]"
+      },
+      {
+        name: "Tashfia Oyshi",
+        image: "src/assets/meettheteam/tashfia-nametag.png",
+        position: "absolute top-[35vh] left-[62vw] w-[16vw]"
+      },
+    ]
+  },
+  Outreach: {
+    title: "Outreach",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+       {
+        type: "image",
+        src: "src/assets/meettheteam/pink-flower.png",
+        position: "absolute top-[5vh] right-[22vw] w-[5vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/director-tag.png",
+        position: "absolute top-[-2vh] left-[25vw] w-[10vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[35vh] right-[27vw] w-[7vw] rotate-270"
+      },
+      { 
+        type: "label",
+        text: "Outreach",
+        position: "absolute bottom-[2vh] left-[2.2vw] w-[16vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[5vh] left-[35vw] w-[7vw] rotate-270"
+      },
+    ],
+    members: [
+      {
+        name: "Brianna Nguyen",
+        image: "src/assets/meettheteam/brianna-nametag.png",
+        position: "absolute top-[5vh] left-[16vw] w-[16vw]"
+      },
+      {
+        name: "Rosa Sakhpour",
+        image: "src/assets/meettheteam/rosa-nametag.png",
+        position: "absolute top-[5vh] left-[36vw] w-[16vw]"
+      },
+      {
+        name: "Tasfia Afrida",
+        image: "src/assets/meettheteam/tasfia-nametag.png",
+        position: "absolute top-[5vh] left-[56vw] w-[16vw]"
+      },
+{
+        name: "Raazia Shohab",
+        image: "src/assets/meettheteam/raazia-nametag.png",
+        position: "absolute top-[35vh] left-[26vw] w-[16vw]"
+      },
+      {
+        name: "Tithi Patel",
+        image: "src/assets/meettheteam/tithi-nametag.png",
+        position: "absolute top-[35vh] left-[44vw] w-[16vw]"
+      },
+      {
+        name: "Niyoosha Tahimi",
+        image: "src/assets/meettheteam/niyoosha-nametag.png",
+        position: "absolute top-[35vh] left-[62vw] w-[16vw]"
+      },
+    ]
+  },
+  CommunityEngagement: {
+    title: "Community Engagement",
+    slideConfig: {
+      width: "105vw",
+      height: "72vh",
+      bgColor: "#A7748E"
+    },
+    decorations: [
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/director-tag.png",
+        position: "absolute top-[-2vh] left-[28vw] w-[10vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/cookies.png",
+        position: "absolute top-[35vh] left-[17vw] w-[12vw]"
+      },
+      {
+        type: "image", 
+        src: "src/assets/meettheteam/beige-flowers.png",
+        position: "absolute top-[5vh] right-[20vw] w-[7vw] rotate-270"
+      },
+      { 
+        type: "label",
+        text: "Community Engagement",
+        position: "absolute bottom-[2vh] left-[7vw] w-[29vw] h-[5.8vh] rounded-[1.64vw] bg-[#270D43] flex justify-center items-center"
+      },
+    ],
+    members: [
+      {
+        name: "Rodiat Adeoye",
+        image: "src/assets/meettheteam/rodiat-nametag.png",
+        position: "absolute top-[5vh] left-[19vw] w-[16vw]"
+      },
+      {
+        name: "Bani Kaur",
+        image: "src/assets/meettheteam/bani-nametag.png",
+        position: "absolute top-[5vh] left-[39vw] w-[16vw]"
+      },
+      {
+        name: "Vasudha Rampal",
+        image: "src/assets/meettheteam/vasudha-nametag.png",
+        position: "absolute top-[5vh] left-[59vw] w-[16vw]"
+      },
+{
+        name: "Baran Parsi",
+        image: "src/assets/meettheteam/baran-nametag.png",
+        position: "absolute top-[35vh] left-[29vw] w-[16vw]"
+      },
+      {
+        name: "Varun Patel",
+        image: "src/assets/meettheteam/varun-nametag.png",
+        position: "absolute top-[35vh] left-[49vw] w-[16vw]"
+      },
+      {
+        name: "Khadija Saleem",
+        image: "src/assets/meettheteam/khadija-nametag.png",
+        position: "absolute top-[35vh] left-[69vw] w-[16vw]"
+      },
+    ]
+  },
+};
+
+const slides = [
+  slideData.coChairs,
+  slideData.Logistics1,
+  slideData.Logistics2,
+  slideData.IT1,
+  slideData.IT2,
+  slideData.Marketing,
+  slideData.Design,
+  slideData.Outreach,
+  slideData.CommunityEngagement
+];
+
+export { slideData, slides };
